Remove src/plugins/vis_default_editor -> src/plugins/visualizations cyclic dependencies (#86988)
* Remove src/plugins/vis_default_editor -> src/plugins/visualizations cyclic dependencies

Part of #86758

* Update agg_common_props.ts

* Schemas -> visualizations

* fix JEST

* make Schemas class private for visualizations plugin

* fix type error

* fix markdown app in visualize app

* remove visualizations from kibana.json's
