+++ src/plugins/region_map/kibana.json
@@ -15,7 +15,7 @@
   ],
   "requiredBundles": [
     "kibanaUtils",
-    "charts", 
+    "charts",
     "visDefaultEditor"
   ]
 }
