+++ src/plugins/vis_default_editor/public/components/agg_common_props.ts
@@ -17,10 +17,10 @@
  * under the License.
  */
 
-import { VisParams } from 'src/plugins/visualizations/public';
-import { IAggType, IAggConfig, AggGroupName } from 'src/plugins/data/public';
-import { Schema } from '../schemas';
-import { EditorVisState } from './sidebar/state/reducers';
+import type { VisParams, Schema } from 'src/plugins/visualizations/public';
+import type { IAggType, IAggConfig, AggGroupName } from 'src/plugins/data/public';
+
+import type { EditorVisState } from './sidebar/state/reducers';
 
 type AggId = IAggConfig['id'];
 type AggParams = IAggConfig['params'];
