+++ src/plugins/vis_default_editor/public/components/agg_group.test.tsx
@@ -20,21 +20,28 @@
 import React from 'react';
 import { mount, shallow } from 'enzyme';
 import { act } from 'react-dom/test-utils';
-import { IAggConfigs, IAggConfig } from 'src/plugins/data/public';
+import type { IAggConfigs, IAggConfig } from 'src/plugins/data/public';
+import { ISchemas } from 'src/plugins/visualizations/public';
+import { createMockedVisEditorSchemas } from 'src/plugins/visualizations/public/mocks';
+
 import { DefaultEditorAggGroup, DefaultEditorAggGroupProps } from './agg_group';
 import { DefaultEditorAgg } from './agg';
 import { DefaultEditorAggAdd } from './agg_add';
-import { ISchemas, Schemas } from '../schemas';
-import { EditorVisState } from './sidebar/state/reducers';
+import type { EditorVisState } from './sidebar/state/reducers';
 
-jest.mock('@elastic/eui', () => ({
-  EuiTitle: 'eui-title',
-  EuiDragDropContext: 'eui-drag-drop-context',
-  EuiDroppable: 'eui-droppable',
-  EuiDraggable: (props: any) => props.children({ dragHandleProps: {} }),
-  EuiSpacer: 'eui-spacer',
-  EuiPanel: 'eui-panel',
-}));
+jest.mock('@elastic/eui', () => {
+  const original = jest.requireActual('@elastic/eui');
+
+  return {
+    ...original,
+    EuiTitle: 'eui-title',
+    EuiDragDropContext: 'eui-drag-drop-context',
+    EuiDroppable: 'eui-droppable',
+    EuiDraggable: (props: any) => props.children({ dragHandleProps: {} }),
+    EuiSpacer: 'eui-spacer',
+    EuiPanel: 'eui-panel',
+  };
+});
 
 jest.mock('./agg', () => ({
   DefaultEditorAgg: () => <div />,
@@ -56,7 +63,7 @@ describe('DefaultEditorAgg component', () => {
     setTouched = jest.fn();
     setValidity = jest.fn();
     reorderAggs = jest.fn();
-    schemas = new Schemas([
+    schemas = createMockedVisEditorSchemas([
       {
         name: 'metrics',
         group: 'metrics',
