+++ src/plugins/vis_default_editor/public/components/agg_group.tsx
@@ -31,6 +31,7 @@ import {
 import { i18n } from '@kbn/i18n';
 
 import { AggGroupNames, AggGroupLabels, IAggConfig, TimeRange } from '../../../data/public';
+import type { Schema } from '../../../visualizations/public';
 import { DefaultEditorAgg } from './agg';
 import { DefaultEditorAggAdd } from './agg_add';
 import { AddSchema, ReorderAggs, DefaultEditorAggCommonProps } from './agg_common_props';
@@ -41,7 +42,6 @@ import {
   getEnabledMetricAggsCount,
 } from './agg_group_helper';
 import { aggGroupReducer, initAggsState, AGGS_ACTION_KEYS } from './agg_group_state';
-import { Schema } from '../schemas';
 
 export interface DefaultEditorAggGroupProps extends DefaultEditorAggCommonProps {
   schemas: Schema[];
