+++ src/plugins/vis_default_editor/public/components/agg_params.tsx
@@ -23,6 +23,7 @@ import { i18n } from '@kbn/i18n';
 import useUnmount from 'react-use/lib/useUnmount';
 
 import { IAggConfig, IndexPattern, AggGroupNames } from '../../../data/public';
+import type { Schema } from '../../../visualizations/public';
 
 import { DefaultEditorAggSelect } from './agg_select';
 import { DefaultEditorAggParam } from './agg_param';
@@ -39,7 +40,7 @@ import {
 } from './agg_params_state';
 import { DefaultEditorCommonProps } from './agg_common_props';
 import { EditorParamConfig, TimeIntervalParam, FixedParam, getEditorConfig } from './utils';
-import { Schema, getSchemaByName } from '../schemas';
+import { getSchemaByName } from '../schemas';
 import { useKibana } from '../../../kibana_react/public';
 import { VisDefaultEditorKibanaServices } from '../types';
