+++ src/plugins/vis_default_editor/public/components/sidebar/state/actions.ts
@@ -17,10 +17,10 @@
  * under the License.
  */
 
-import { Vis, VisParams } from 'src/plugins/visualizations/public';
-import { IAggConfig } from 'src/plugins/data/public';
+import { Vis, VisParams, Schema } from 'src/plugins/visualizations/public';
+import type { IAggConfig } from 'src/plugins/data/public';
+
 import { EditorStateActionTypes } from './constants';
-import { Schema } from '../../../schemas';
 
 export interface ActionType<T, P> {
   type: T;
