+++ src/plugins/vis_type_metric/public/metric_vis_type.ts
@@ -18,11 +18,10 @@
  */
 
 import { i18n } from '@kbn/i18n';
-import { BaseVisTypeOptions } from 'src/plugins/visualizations/public';
 import { MetricVisOptions } from './components/metric_vis_options';
 import { ColorSchemas, colorSchemas, ColorMode } from '../../charts/public';
+import { BaseVisTypeOptions } from '../../visualizations/public';
 import { AggGroupNames } from '../../data/public';
-import { Schemas } from '../../vis_default_editor/public';
 import { toExpressionAst } from './to_ast';
 
 export const createMetricVisTypeDefinition = (): BaseVisTypeOptions => ({
@@ -83,7 +82,7 @@ export const createMetricVisTypeDefinition = (): BaseVisTypeOptions => ({
       colorSchemas,
     },
     optionsTemplate: MetricVisOptions,
-    schemas: new Schemas([
+    schemas: [
       {
         group: AggGroupNames.Metrics,
         name: 'metric',
@@ -120,6 +119,6 @@ export const createMetricVisTypeDefinition = (): BaseVisTypeOptions => ({
         max: 1,
         aggFilter: ['!geohash_grid', '!geotile_grid', '!filter'],
       },
-    ]),
+    ],
   },
 });
