+++ src/plugins/vis_type_table/public/table_vis_type.ts
@@ -18,7 +18,6 @@
  */
 import { i18n } from '@kbn/i18n';
 import { AggGroupNames } from '../../data/public';
-import { Schemas } from '../../vis_default_editor/public';
 import { BaseVisTypeOptions } from '../../visualizations/public';
 
 import { TableOptions } from './components/table_vis_options_lazy';
@@ -51,7 +50,7 @@ export const tableVisTypeDefinition: BaseVisTypeOptions<TableVisParams> = {
   },
   editorConfig: {
     optionsTemplate: TableOptions,
-    schemas: new Schemas([
+    schemas: [
       {
         group: AggGroupNames.Metrics,
         name: 'metric',
@@ -85,7 +84,7 @@ export const tableVisTypeDefinition: BaseVisTypeOptions<TableVisParams> = {
         max: 1,
         aggFilter: ['!filter'],
       },
-    ]),
+    ],
   },
   toExpressionAst,
   hierarchicalData: (vis) => vis.params.showPartialRows || vis.params.showMetricsAtAllLevels,
