+++ src/plugins/vis_type_vislib/public/heatmap.ts
@@ -20,7 +20,7 @@
 import { i18n } from '@kbn/i18n';
 import { Position } from '@elastic/charts';
 
-import { RangeValues, Schemas } from '../../vis_default_editor/public';
+import { RangeValues } from '../../vis_default_editor/public';
 import { AggGroupNames } from '../../data/public';
 import { ColorSchemas, ColorSchemaParams } from '../../charts/public';
 import { VIS_EVENT_TO_TRIGGER, BaseVisTypeOptions } from '../../visualizations/public';
@@ -88,7 +88,7 @@ export const heatmapVisTypeDefinition: BaseVisTypeOptions<BasicVislibParams> = {
   editorConfig: {
     collections: getHeatmapCollections(),
     optionsTemplate: HeatmapOptions,
-    schemas: new Schemas([
+    schemas: [
       {
         group: AggGroupNames.Metrics,
         name: 'metric',
@@ -136,6 +136,6 @@ export const heatmapVisTypeDefinition: BaseVisTypeOptions<BasicVislibParams> = {
         max: 1,
         aggFilter: ['!geohash_grid', '!geotile_grid', '!filter'],
       },
-    ]),
+    ],
   },
 };
