+++ src/plugins/vis_type_xy/public/vis_types/area.tsx
@@ -24,7 +24,6 @@ import { i18n } from '@kbn/i18n';
 import { euiPaletteColorBlind } from '@elastic/eui/lib/services';
 import { Fit, Position } from '@elastic/charts';
 
-import { Schemas } from '../../../vis_default_editor/public';
 import { AggGroupNames } from '../../../data/public';
 import { VIS_EVENT_TO_TRIGGER } from '../../../visualizations/public';
 import { defaultCountLabel, LabelRotation } from '../../../charts/public';
@@ -139,7 +138,7 @@ export const getAreaVisTypeDefinition = (
   editorConfig: {
     collections: getConfigCollections(),
     optionTabs: getOptionTabs(showElasticChartsOptions),
-    schemas: new Schemas([
+    schemas: [
       {
         group: AggGroupNames.Metrics,
         name: 'metric',
@@ -196,6 +195,6 @@ export const getAreaVisTypeDefinition = (
           tooltip: <SplitTooltip />,
         }),
       },
-    ]),
+    ],
   },
 });
