+++ src/plugins/visualizations/public/index.ts
@@ -37,7 +37,14 @@ export { getSchemas as getVisSchemas } from './legacy/build_pipeline';
 /** @public types */
 export { VisualizationsSetup, VisualizationsStart };
 export { VisGroups } from './vis_types';
-export type { VisTypeAlias, VisType, BaseVisTypeOptions, ReactVisTypeOptions } from './vis_types';
+export type {
+  VisTypeAlias,
+  VisType,
+  BaseVisTypeOptions,
+  ReactVisTypeOptions,
+  Schema,
+  ISchemas,
+} from './vis_types';
 export { VisParams, SerializedVis, SerializedVisData, VisData } from './vis';
 export type VisualizeEmbeddableFactoryContract = PublicContract<VisualizeEmbeddableFactory>;
 export type VisualizeEmbeddableContract = PublicContract<VisualizeEmbeddable>;
