+++ src/plugins/visualizations/public/vis_types/base_vis_type.ts
@@ -18,9 +18,10 @@
  */
 
 import { defaultsDeep } from 'lodash';
-import { ISchemas } from 'src/plugins/vis_default_editor/public';
+
 import { VisParams } from '../types';
 import { VisType, VisTypeOptions, VisGroups } from './types';
+import { Schemas } from './schemas';
 
 interface CommonBaseVisTypeOptions<TVisParams>
   extends Pick<
@@ -102,6 +103,7 @@ export class BaseVisType<TVisParams = VisParams> implements VisType<TVisParams> 
   public readonly inspectorAdapters;
   public readonly toExpressionAst;
   public readonly getInfoMessage;
+  public readonly schemas;
 
   constructor(opts: BaseVisTypeOptions<TVisParams>) {
     if (!opts.icon && !opts.image) {
@@ -133,10 +135,8 @@ export class BaseVisType<TVisParams = VisParams> implements VisType<TVisParams> 
     this.inspectorAdapters = opts.inspectorAdapters;
     this.toExpressionAst = opts.toExpressionAst;
     this.getInfoMessage = opts.getInfoMessage;
-  }
 
-  public get schemas(): ISchemas {
-    return this.editorConfig?.schemas ?? [];
+    this.schemas = new Schemas(this.editorConfig?.schemas ?? []);
   }
 
   public get requiresSearch(): boolean {
