+++ src/plugins/visualizations/public/vis_types/index.ts
@@ -18,7 +18,8 @@
  */
 
 export * from './types_service';
+export { Schemas } from './schemas';
 export { VisGroups } from './types';
-export type { VisType } from './types';
+export type { VisType, ISchemas, Schema } from './types';
 export type { BaseVisTypeOptions } from './base_vis_type';
 export type { ReactVisTypeOptions } from './react_vis_type';
